requirements.txt
flask==3.0.3
flask-wtf==1.2.1
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
supabase==2.5.1
flask-login==0.6.3
bcrypt==4.2.0
pillow==10.4.0
gunicorn==23.0.0
